+++ lib/tests/step.func.spec.js
@@ -1,0 +1,46 @@
+var chai = require('chai')
+var sinon = require('sinon')
+
+var expect = chai.expect
+chai.use(require('sinon-chai'))
+chai.use(require('chai-properties'))
+
+var controllers = require('../autoBlock')
+
+describe('step .func', () => {
+  var sandbox
+
+  beforeEach(() => {
+    sandbox = sinon.sandbox.create()
+  })
+
+  afterEach(() => {
+    sandbox.restore()
+  })
+
+  it('should survive throw', (done) => {
+    var one = (results, cb) => {
+      throw new Error('HES DEAD JIM')
+    }
+
+    var final = (err, response) => {
+      expect(err).to.exist
+      done()
+    }
+
+    var controller = {
+      data: {
+        event: { foo: 'bar' }
+      },
+      done: final
+    }
+
+    controller.block = {
+      'one': {
+        func: one
+      }
+    }
+
+    controllers.run(controller)
+  })
+})
